+++ NEWS.md
@@ -997,6 +997,18 @@ Deprecated or removed
 
   * `gradient` is deprecated and will be removed in the next release ([#23816]).
 
+  * `lu(A::AbstractMatrix[, pivot])` has been deprecated in favor of `lufact(A[, pivot])`.
+    Whereas `lu(A[, pivot])` returns a tuple of arrays, `lufact(A[, pivot])` returns an `LU` object.
+    So for a direct replacement, use `(lufact(A[, pivot])...,)`. But going forward, consider using
+    the direct result of `lufact(A[, pivot])` instead, either destructured into its components
+    (`l, u, p = lufact(A[, pivot])`) or as an `LU` object (`lup = lufact(A)`) ([#26997]).
+
+  * `lu(x::Number)` has been deprecated in favor of `lufact(x::Number)`.
+    Whereas `lu(x::Number)` returns a tuple of numbers, `lufact(x::Number)` returns a tuple of arrays
+    for consistency with other `lufact` methods. So for a direct replacement, use `first.((lufact(x)...,))`.
+    But going forward, consider using the direct result of `lufact(x)` instead, either destructured
+    into its components (`l, u, p = lufact(x)`) or as an `LU` object (`lup = lufact(x)`) ([#26997]).
+
   * The timing functions `tic`, `toc`, and `toq` are deprecated in favor of `@time` and `@elapsed`
     ([#17046]).